requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
pandas==2.1.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydruid==0.6.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
